Add 'owner' privilege to client resource to creator
This is a tiny first step towards ACL support. When a new 'client' gets
created in tt-api, the controller will now automatically add mark the
current principal(user) as the owner.

This works, but needs some refactoring because it's a lot of code for a
controller and we really wanna centralize it.