Show ACL on every resource if you're an admin.
Just had this idea this afternoon: wouldn't it be nice if we can quickly
see the ACL list of every resource in an API by a click of the button.

This middleware does that. If you're an admin, a new button gets
rendered everywhere.

Click it, and you'll see all the relevant privileges as fetched from
a12n-server.